change the order of input parameter
