+++ source/module_io/read_input.h
@@ -1,10 +1,10 @@
 #ifndef READ_INPUT_H
 #define READ_INPUT_H
 
-#include <string>
-
 #include "input_item.h"
 #include "module_parameter/parameter.h"
+
+#include <string>
 
 namespace ModuleIO
 {
@@ -78,21 +78,37 @@ class ReadInput
     void add_item(const Input_Item& item);
     // add bcast functions for global values
     void set_globalv_bcast();
-    // general items
-    void item_general();
-    // items for pw
-    void item_pw();
-    // items for sdft
-    void item_sdft();
-    // items for relax
-    void item_relax();
+    // system items
+    void item_system();
+    // items for electronic structure
+    void item_elec_stru();
     // items for lcao
     void item_lcao();
-    // items for postprocess
-    void item_postprocess();
+    // items for relax
+    void item_relax();
     // items for md
     void item_md();
-    // items for others
+    // items for ofdft
+    void item_ofdft();
+    // items for sdft
+    void item_sdft();
+    // items for deepks
+    void item_deepks();
+    // items for real time tddft
+    void item_rt_tddft();
+    // items for linear response tddft
+    void item_lr_tddft();
+    // items for output
+    void item_output();
+    // items for postprocess
+    void item_postprocess();
+    // items for some models
+    void item_model();
+    // items for exx
+    void item_exx();
+    // items for dft+u
+    void item_dftu();
+    // items for other
     void item_others();
 
   private:
