+++ source/module_io/test_serial/CMakeLists.txt
@@ -6,14 +6,19 @@ remove_definitions(-D__MPI)
 add_library(
     io_input_serial
     OBJECT
-    ../read_input_item_other.cpp
-    ../read_input_item_general.cpp
-    ../read_input_item_pw.cpp
-    ../read_input_item_sdft.cpp
+    ../read_input_item_system.cpp
+    ../read_input_item_elec_stru.cpp
     ../read_input_item_relax.cpp
-    ../read_input_item_lcao.cpp
-    ../read_input_item_postprocess.cpp
     ../read_input_item_md.cpp
+    ../read_input_item_ofdft.cpp
+    ../read_input_item_sdft.cpp
+    ../read_input_item_tddft.cpp
+    ../read_input_item_deepks.cpp
+    ../read_input_item_model.cpp
+    ../read_input_item_postprocess.cpp  
+    ../read_input_item_exx_dftu.cpp
+    ../read_input_item_other.cpp   
+    ../read_input_item_output.cpp
     ../read_input.cpp
     ../bcast_globalv.cpp
 )
